Allow timeout option for WinRM commands
Allows end users (e.g. InSpec test coders) to specify a timeout for a potentially long running command.
If the timeout is reached, the command is expected to be terminated on the host and an exception is raised (a subsequent change to InSpec will handle this exception).
This complements recent changes to the base connection and ssh connection in train: https://github.com/inspec/train/pull/625

Signed-off-by: James Stocks <jstocks@chef.io>